requirements.txt
torch>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
safetensors>=0.3.0
sentencepiece>=0.1.99
accelerate>=0.21.0
bitsandbytes>=0.40.0
einops>=0.6.1
mediapy>=1.1.0
huggingface_hub>=0.16.0
Pillow>=9.5.0

